requirements.txt
# Document Processing
docling>=2.0.0
docling-core>=2.0.0

# RAG Framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0

# Vector Store (Local)
chromadb>=0.5.0

# Embeddings
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# FastAPI (Web API)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
